refactor(cohort-analyzer): normalize module size and split large files
Keep Cohort Analyzer sources roughly under 500 lines per file by extracting focused modules instead of monolithic components.

- Split cohortAnalyzerStyling into core, layout, and summary style modules
- Split histogram panel styled pieces into core and modal/survival modules
- Extract Histogram hooks: panel bootstrap, strip DnD, resize handlers, beside metrics, survival beside-Venn card style
- Extract Histogram UI: strip chart row, beside-Venn histogram portal, survival layout fragments
- Split expanded chart modal into header plus survival, Venn, and histogram tab bodies
- Add histogram modal survival downloads and KM-derived helpers; summary view shell for the analyzer
- Slim CohortAnalyzer by consolidating useCohortAnalyzer destructuring

Made-with: Cursor
