+++ src/pages/CohortAnalyzer/cohortAnalyzerStyling/summaryStyles.js
@@ -1,0 +1,318 @@
+import {
+  CA_SURVIVAL_CARD_MIN_WIDTH,
+  CA_SURVIVAL_CARD_MIN_HEIGHT,
+} from '../store/cohortAnalyzerLayoutConstants';
+
+/** Chart summary view: Venn row, tabs, controls, category cards, table area. */
+export const cohortAnalyzerSummaryStyles = (theme) => ({
+  chartSummaryMain: {
+    width: '100%',
+    marginBottom: 45,
+    marginTop: '16px',
+  },
+  vennSurvivalRow: {
+    display: 'flex',
+    flexDirection: 'row',
+    flexWrap: 'wrap',
+    alignItems: 'flex-start',
+    alignContent: 'flex-start',
+    gap: 26,
+    width: '100%',
+    minWidth: 0,
+    overflowX: 'visible',
+    overflowY: 'visible',
+    [theme.breakpoints.down('md')]: {
+      flexDirection: 'column',
+      flexWrap: 'nowrap',
+      gap: 20,
+    },
+  },
+  vennColumn: {
+    flex: '1 1 0',
+    minWidth: 0,
+    maxWidth: '100%',
+    display: 'flex',
+    flexDirection: 'column',
+    alignItems: 'stretch',
+  },
+  survivalBesideVennColumn: {
+    flex: '1 1 0',
+    minWidth: CA_SURVIVAL_CARD_MIN_WIDTH,
+    minHeight: CA_SURVIVAL_CARD_MIN_HEIGHT,
+    maxWidth: '100%',
+    display: 'flex',
+    flexDirection: 'column',
+    alignItems: 'stretch',
+    alignSelf: 'stretch',
+  },
+  rightSideAnalyzerInnerContainer: {
+    display: 'flex',
+    marginBottom: 0,
+    flexDirection: 'column',
+    width: '100%',
+    minWidth: 0,
+    flex: '1 1 auto',
+    alignItems: 'stretch',
+  },
+  rightSideAnalyzerHeader: {
+    display: 'flex',
+    justifyContent: 'space-between',
+    position: 'relative',
+    width: '100%',
+    alignSelf: 'center',
+    alignItems: 'center',
+    borderBottom: '2.5px solid #679AAA',
+    maxHeight: '70px',
+    '& h1': {
+      fontFamily: 'Poppins',
+      fontSize: '40px',
+      fontWeight: 400,
+      textAlign: 'left',
+      color: '#0D3A3F',
+      marginLeft: '30px',
+      marginTop: '12px',
+      marginBottom: '12px',
+    },
+  },
+  readmeButton: {
+    background: '#3E6069',
+    color: '#FFFFFF',
+    border: '1px solid #2A4951',
+    borderRadius: 4,
+    fontFamily: 'Poppins',
+    fontSize: 11,
+    fontWeight: 600,
+    height: 30,
+    minWidth: 115,
+    marginRight: 30,
+    cursor: 'pointer',
+    letterSpacing: '0.02em',
+  },
+  summaryViewShell: {
+    display: 'flex',
+    flexDirection: 'column',
+    alignSelf: 'stretch',
+    width: '100%',
+    flex: 1,
+    minHeight: 0,
+    minWidth: 0,
+    paddingTop: 14,
+    boxSizing: 'border-box',
+  },
+  summaryTabs: {
+    display: 'flex',
+    flexDirection: 'row',
+    flexWrap: 'wrap',
+    alignItems: 'stretch',
+    justifyContent: 'flex-start',
+    width: '100%',
+    boxSizing: 'border-box',
+    gap: 8,
+    paddingLeft: 30,
+    paddingRight: 30,
+    borderBottom: '1px solid #BFC9CD',
+    marginBottom: 0,
+  },
+  summaryTab: {
+    flex: '0 0 auto',
+    width: 'auto',
+    border: '1px solid #BCBCBC',
+    borderBottom: 'none',
+    background: '#E8E8E8',
+    color: '#1C2B33',
+    fontFamily: 'Poppins',
+    fontSize: 16,
+    padding: '8px 16px',
+    cursor: 'pointer',
+    marginRight: 0,
+    textAlign: 'left',
+    whiteSpace: 'nowrap',
+  },
+  summaryTabActive: {
+    background: '#FFFFFF',
+    color: '#000000',
+    borderTop: '3px solid #3E7D8E',
+    fontWeight: 500,
+  },
+  summaryTabPanel: {
+    flex: 1,
+    minHeight: 0,
+    minWidth: 0,
+    width: '100%',
+    alignSelf: 'stretch',
+    boxSizing: 'border-box',
+    backgroundColor: '#EFF5FB',
+    padding: '14px 30px 24px 30px',
+    marginTop: 0,
+    display: 'flex',
+    flexDirection: 'column',
+  },
+  chartTopControlRow: {
+    alignSelf: 'stretch',
+    width: 'auto',
+    marginLeft: -30,
+    marginRight: -30,
+    paddingLeft: 30,
+    paddingRight: 30,
+    boxSizing: 'border-box',
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'space-between',
+    marginBottom: 14,
+    gap: 10,
+    flexWrap: 'wrap',
+    paddingBottom: 10,
+    borderBottom: '1px solid #4E8191',
+    marginTop: -5,
+  },
+  categoryPills: {
+    display: 'flex',
+    alignItems: 'center',
+    flexWrap: 'wrap',
+    gap: 6,
+  },
+  categoryPillButton: {
+    border: '1px solid #95A4AA',
+    borderRadius: 3,
+    background: '#646464',
+    fontFamily: 'Poppins',
+    color: 'white',
+    fontSize: 10.5,
+    fontWeight: 500,
+    lineHeight: '13px',
+    padding: '4px 8px',
+    letterSpacing: '0.01em',
+    cursor: 'default',
+  },
+  chartActionButtons: {
+    display: 'flex',
+    alignItems: 'center',
+    gap: 8,
+  },
+  addChartButton: {
+    border: '1px solid #295665',
+    borderRadius: 4,
+    background: '#006A8F',
+    color: '#FFFFFF',
+    fontFamily: 'Poppins',
+    fontWeight: 600,
+    fontSize: 12,
+    height: 33,
+    minWidth: 126,
+    cursor: 'default',
+    '& span': {
+      fontSize: 16,
+      marginLeft: 8,
+      position: 'relative',
+      top: 1,
+    },
+  },
+  downloadAllButton: {
+    border: '1px solid #2C4D59',
+    borderRadius: 4,
+    background: '#556469',
+    color: '#FFFFFF',
+    fontFamily: 'Poppins',
+    fontWeight: 600,
+    fontSize: 11.5,
+    height: 33,
+    minWidth: 128,
+    cursor: 'default',
+    letterSpacing: '0.02em',
+  },
+  rightSideAnalyzerHeader2: {
+    display: 'flex',
+    justifyContent: 'space-between',
+    position: 'relative',
+    minWidth: '30%',
+    alignItems: 'flex-start',
+    paddingBottom: 0,
+    paddingTop: 0,
+    backgroundColor: 'red',
+    zIndex: 900,
+    '& p': {
+      fontFamily: 'Open Sans',
+      fontSize: '15px',
+      paddingBottom: 0,
+      width: '100%',
+      lineHeight: '20.8px',
+      textAlign: 'left',
+      color: 'black',
+      margin: 1,
+      marginTop: 0,
+      marginLeft: '20px',
+      [theme.breakpoints.down('lg')]: {
+        width: '100%',
+        marginLeft: 0,
+      },
+    },
+  },
+  alert: {
+    position: 'absolute',
+    top: 20,
+    margin: 'auto',
+  },
+  catagoryCard: {
+    borderRadius: '8px',
+    display: 'flex',
+    flexDirection: 'column',
+    justifyContent: 'center',
+    minWidth: 297,
+    height: 194,
+    marginLeft: 0,
+    paddingBottom: 30,
+    border: '1px solid #8B98AF',
+    alignItems: 'center',
+    alignContent: 'center',
+    '& h3': {
+      fontFamily: 'Poppins',
+      fontSize: '18px',
+      fontWeight: 400,
+      textAlign: 'center',
+      color: '#000',
+    },
+  },
+  catagoryCardChildren: {
+    width: '50%',
+    minHeight: 75,
+    alignSelf: 'center',
+    display: 'flex',
+    overflow: 'hidden',
+    flexDirection: 'column',
+    justifyContent: 'center',
+    alignItems: 'flex-start',
+    color: '#5D7B87',
+    '& p': {
+      top: 325,
+      gap: 9,
+      fontSize: 14,
+      fontFamily: 'Poppins',
+      fontWeight: 300,
+      lineHeight: 2,
+      margin: 0,
+      cursor: 'pointer',
+      '& input': {
+        cursor: 'pointer',
+      },
+    },
+  },
+  rightSideTableContainer: {
+    width: '100%',
+    maxHeight: 540,
+    overflowY: 'scroll',
+    borderTop: '3px #679AAA solid',
+    '&::-webkit-scrollbar': {
+      width: '6px',
+    },
+    '&::-webkit-scrollbar-thumb': {
+      width: '6px',
+      backgroundColor: '#003F74',
+    },
+    '&::-webkit-scrollbar-track': {
+      background: '#CECECE',
+    },
+    [theme.breakpoints.down('lg')]: {
+      width: '100%',
+    },
+  },
+});
